requirements.txt
# Core data handling
numpy==2.4.3
pandas==2.3.3
pillow==12.1.1
requests==2.32.5
tqdm==4.67.3
torch
datasets
safetensors

# Evaluation / analysis
scikit-learn==1.8.0
scipy==1.17.1

# Geospatial pipeline
geopandas==1.1.3
shapely==2.1.2
pyproj==3.7.2
rasterio==1.5.0
earthengine-api==1.7.17
geemap==0.37.1

# Local LoRA fine-tuning on Apple Silicon
mlx; platform_system == "Darwin" and platform_machine == "arm64"
mlx-vlm; platform_system == "Darwin" and platform_machine == "arm64"
